Implement fallback handler logic
