+++ contracts/base/ModuleManager.sol
@@ -16,7 +16,7 @@ contract ModuleManager is SelfAuthorized, Executor {
     address public constant SENTINEL_MODULES = address(0x1);
 
     mapping (address => address) internal modules;
-    
+
     function setupModules(address to, bytes memory data)
         internal
     {
