+++ contracts/interfaces/ERC777TokensRecipient.sol
@@ -1,0 +1,12 @@
+pragma solidity ^0.5.0;
+
+interface ERC777TokensRecipient {
+    function tokensReceived(
+        address operator,
+        address from,
+        address to,
+        uint256 amount,
+        bytes calldata data,
+        bytes calldata operatorData
+    ) external;
+}
