+++ package.json
@@ -51,7 +51,7 @@
   },
   "dependencies": {
     "openzeppelin-solidity": "^2.0.0",
-    "solc": "^0.5.0",
+    "solc": "^0.5.9",
     "truffle-hdwallet-provider": "0.0.7-beta.1"
   }
 }
