+++ test/gnosisSafeDeploymentViaTx.js
@@ -1,77 +1,107 @@
 const utils = require('./utils/general')
 const safeUtils = require('./utils/execution')
+const fs = require('fs')
+const randomBuffer = require("random-buffer")
 const ethUtil = require('ethereumjs-util')
-const abi = require('ethereumjs-abi')
+const EthereumTx = require('ethereumjs-tx')
 
 const GnosisSafe = artifacts.require("./GnosisSafe.sol")
-const Proxy = artifacts.require("./Proxy.sol")
-const ProxyFactory = artifacts.require("./ProxyFactory.sol")
 const MockContract = artifacts.require('./MockContract.sol');
 const MockToken = artifacts.require('./Token.sol');
 
-contract('GnosisSafe via create2', function(accounts) {
+contract('GnosisSafe Trustless Deployment', function(accounts) {
 
     const CALL = 0
 
     const gasPrice = web3.toWei('20', 'gwei')
+    const payingProxyJson = JSON.parse(fs.readFileSync("./build/contracts/PayingProxy.json"))
+    const PayingProxy = web3.eth.contract(payingProxyJson.abi)
 
     let funder
-    let proxyFactory
     let gnosisSafeMasterCopy
     let lw
 
-    let getCreationData = async function(gasToken, userCosts, creationNonce) {
-        gnosisSafeData = await gnosisSafeMasterCopy.contract.setup.getData([lw.accounts[0], lw.accounts[1], lw.accounts[2]], 2, 0, "0x", gasToken, userCosts, 0)
+    let setSignature = function(tx) {
+        while(true) {
+            try {
+              tx.s = randomBuffer(32) // Could be provided by client
+              tx.r = randomBuffer(32) // Could be provided by server
+              setV(tx)
+              return // Success return
+            } catch (e) {
+            }
+        }
+        throw "No valid signature found"
+    }
 
-        let proxyCreationCode = await proxyFactory.proxyCreationCode()
-        assert.equal(proxyCreationCode, Proxy.bytecode)
-        let constructorData = abi.rawEncode(
-            ['address'], 
-            [ gnosisSafeMasterCopy.address ]
-        ).toString('hex')
-        let encodedNonce = abi.rawEncode(['uint256'], [creationNonce]).toString('hex')
-        let target = "0x" + ethUtil.generateAddress2(proxyFactory.address, ethUtil.keccak256("0x" + ethUtil.keccak256(gnosisSafeData).toString("hex") + encodedNonce), proxyCreationCode + constructorData).toString("hex")
+    let setV = function(tx) {
+        for(v = 27; v <= 30; v++) {
+            try {
+              tx.v = ethUtil.intToBuffer(v)
+              tx.getSenderAddress()
+              return
+            } catch (e) {
+            }
+        }
+        throw "No valid v found"
+    }
+
+    let getCreationData = async function(gasToken, userCosts, gasLimit) {
+        gnosisSafeData = await gnosisSafeMasterCopy.contract.setup.getData([lw.accounts[0], lw.accounts[1], lw.accounts[2]], 2, 0, "0x", 0, 0, 0, 0)
+
+        let rawTx = {
+            value: 0,
+            data: PayingProxy.new.getData(gnosisSafeMasterCopy.address, gnosisSafeData, funder, gasToken, userCosts, {data: payingProxyJson.bytecode}),
+        }
+
+        var tx = new EthereumTx(rawTx);
+        tx.nonce = 0
+        tx.gasPrice = web3.toHex(gasPrice)
+        tx.gasLimit = gasLimit || 500000
+        setSignature(tx)
+
+        let sender = tx.getSenderAddress().toString("hex")
+        console.log("    Random sender: 0x" + sender)
+        let target = "0x" + ethUtil.generateAddress(tx.getSenderAddress(), 0).toString("hex")
         console.log("    Predicted safe address: " + target)
         assert.equal(await web3.eth.getCode(target), "0x")
         return {
+            sender: sender,
             safe: target,
-            data: gnosisSafeData,
-            gasToken: gasToken,
+            tx: tx,
             userCosts: userCosts,
-            gasPrice: gasPrice,
-            creationNonce: creationNonce
+            gasPrice: gasPrice
         }
     }
 
     let deployWithCreationData = async function(creationData) {
-        let tx = proxyFactory.createProxyWithNonce(
-            gnosisSafeMasterCopy.address, creationData.data, creationData.creationNonce, 
-            {from: funder, gasPrice: gasPrice}
-        )
+        // Gnosis funds sender
+        await web3.eth.sendTransaction({from: funder, to: creationData.sender, value: creationData.tx.getUpfrontCost(), gasPrice: creationData.gasPrice})
 
-        console.log("    Deployed safe address: " + utils.getParamFromTxEvent(
-            await tx,
-            'ProxyCreation', 'proxy', proxyFactory.address, GnosisSafe, 'create Gnosis Safe',
-        ).address)
+        var raw = '0x' + creationData.tx.serialize().toString('hex')
+        utils.logGasUsage("deploy safe", await web3.eth.getTransactionReceipt(await web3.eth.sendRawTransaction(raw)))
     }
 
     beforeEach(async function () {
-        proxyFactory = await ProxyFactory.new()
         funder = accounts[5]
         // Create lightwallet
         lw = await utils.createLightwallet()
-        gnosisSafeMasterCopy = await GnosisSafe.new()
-        gnosisSafeMasterCopy.setup([accounts[0]], 1, 0, "0x", 0, 0, 0)
+        gnosisSafeMasterCopy = await utils.deployContract("deploying Gnosis Safe Mastercopy", GnosisSafe)
+        gnosisSafeMasterCopy.setup([accounts[0]], 1, 0, "0x", 0, 0, 0, 0)
     })
 
     it('should create safe from random account and pay in ETH', async () => {
 
         // Estimate safe creation costs
-        let gnosisSafeData = await gnosisSafeMasterCopy.contract.setup.getData([lw.accounts[0], lw.accounts[1], lw.accounts[2]], 2, 0, "0x", 0, 0, 0)
-        
-        let creationNonce = new Date().getTime()
-        let estimate = (await proxyFactory.createProxyWithNonce.estimateGas(gnosisSafeMasterCopy.address, gnosisSafeData, creationNonce)) + 9000
-        let creationData = await getCreationData(0, estimate * gasPrice, creationNonce)
+        let gnosisSafeData = await gnosisSafeMasterCopy.contract.setup.getData([lw.accounts[0], lw.accounts[1], lw.accounts[2]], 2, 0, "0x", 0, 0, 0, 0)
+
+        let rawTx = {
+            value: 0,
+            data: PayingProxy.new.getData(gnosisSafeMasterCopy.address, gnosisSafeData, funder, 0, 0, { data: payingProxyJson.bytecode }),
+        }
+        let estimate = web3.eth.estimateGas(rawTx) + 80000
+
+        let creationData = await getCreationData(0, (estimate + 21000) * gasPrice, estimate)
 
         // User funds safe
         await web3.eth.sendTransaction({from: accounts[1], to: creationData.safe, value: creationData.userCosts})
@@ -79,13 +109,11 @@ contract('GnosisSafe via create2', function(accounts) {
         let funderBalance = await web3.eth.getBalance(funder).toNumber()
 
         await deployWithCreationData(creationData)
-        assert.equal(await web3.eth.getCode(creationData.safe), await proxyFactory.proxyRuntimeCode())
+        assert.equal(await web3.eth.getCode(creationData.safe), payingProxyJson.deployedBytecode)
 
         let gnosisSafe = GnosisSafe.at(creationData.safe)
         assert.deepEqual(await gnosisSafe.getOwners(), [lw.accounts[0], lw.accounts[1], lw.accounts[2]])
-        let funderDiff = await web3.eth.getBalance(funder) - funderBalance
-        console.log("    Executor earned " + web3.fromWei(funderDiff, 'ether') + " ETH")
-        assert.ok(funderDiff > 0)
+        assert.equal(await web3.eth.getBalance(funder).toNumber(), funderBalance)
 
         await web3.eth.sendTransaction({from: accounts[1], to: gnosisSafe.address, value: web3.toWei(1.1, 'ether')})
         await safeUtils.executeTransaction(lw, gnosisSafe, 'executeTransaction withdraw 0.5 ETH', [lw.accounts[0], lw.accounts[2]], accounts[0], web3.toWei(0.5, 'ether'), "0x", CALL, accounts[8])
@@ -96,10 +124,8 @@ contract('GnosisSafe via create2', function(accounts) {
         // Deploy token
         let token = await safeUtils.deployToken(accounts[0])
 
-        let creationNonce = new Date().getTime()
-
         // We just set an fix amount of tokens to pay
-        let creationData = await getCreationData(token.address, 1337, creationNonce)
+        let creationData = await getCreationData(token.address, 1337)
 
         // User funds safe
         token.transfer(creationData.safe, creationData.userCosts, {from: accounts[0]})
@@ -107,7 +133,7 @@ contract('GnosisSafe via create2', function(accounts) {
         assert.equal(await token.balances(funder), 0)
 
         await deployWithCreationData(creationData)
-        assert.equal(await web3.eth.getCode(creationData.safe), await proxyFactory.proxyRuntimeCode())
+        assert.equal(await web3.eth.getCode(creationData.safe), payingProxyJson.deployedBytecode)
 
         let gnosisSafe = GnosisSafe.at(creationData.safe)
         assert.deepEqual(await gnosisSafe.getOwners(), [lw.accounts[0], lw.accounts[1], lw.accounts[2]])
@@ -124,8 +150,8 @@ contract('GnosisSafe via create2', function(accounts) {
 
     it('should fail if ether payment fails', async () => {
         // We just set an fix amount of eth to pay
-        let creationData = await getCreationData(0, 5000 * gasPrice, new Date().getTime())
-        utils.assertRejects(deployWithCreationData(creationData), "Deployment without enough ETH should fail")
+        let creationData = await getCreationData(0, 500000 * gasPrice)
+        await deployWithCreationData(creationData)
         assert.equal(await web3.eth.getCode(creationData.safe), '0x')
     })
 
@@ -133,37 +159,30 @@ contract('GnosisSafe via create2', function(accounts) {
         // Deploy token
         let token = await safeUtils.deployToken(accounts[0])
 
-        let nonce = new Date().getTime()
         // We just set an fix amount of tokens to pay
-        let creationData = await getCreationData(token.address, 1337, nonce)
-        utils.assertRejects(deployWithCreationData(creationData), "Deployment without enough tokens should fail")
+        let creationData = await getCreationData(token.address, 1337)
+        await deployWithCreationData(creationData)
         assert.equal(await web3.eth.getCode(creationData.safe), '0x')
 
         let mockContract = await MockContract.new();
         let mockToken = MockToken.at(mockContract.address);
 
         await mockContract.givenAnyRunOutOfGas();
-        creationData = await getCreationData(mockToken.address, 1337, nonce)
-        utils.assertRejects(deployWithCreationData(creationData), "Deployment without enough gas should fail")
+        creationData = await getCreationData(mockToken.address, 1337)
+        await deployWithCreationData(creationData);
         assert.equal(await web3.eth.getCode(creationData.safe), '0x');
 
 
         await mockContract.givenAnyRevert();
-        creationData = await getCreationData(mockToken.address, 1337, nonce);
-        utils.assertRejects(deployWithCreationData(creationData), "Deployment when payment fails should fail")
+        creationData = await getCreationData(mockToken.address, 1337);
+        await deployWithCreationData(creationData);
         assert.equal(await web3.eth.getCode(creationData.safe), '0x');
 
 
         await mockContract.givenAnyReturnBool(false);
-        creationData = await getCreationData(mockToken.address, 1337, nonce);
-        utils.assertRejects(deployWithCreationData(creationData), "Deployment when payment transfer returns false should fail")
+        creationData = await getCreationData(mockToken.address, 1337);
+        await deployWithCreationData(creationData);
         assert.equal(await web3.eth.getCode(creationData.safe), '0x');
-
-        // We deploy a proxy to the same predicted address to make sure that even after a failure we could deploy a successfull one
-        await mockContract.givenAnyReturnBool(true);
-        creationData = await getCreationData(mockToken.address, 1337, nonce);
-        await deployWithCreationData(creationData)
-        assert.equal(await web3.eth.getCode(creationData.safe), await proxyFactory.proxyRuntimeCode())
 
     });
 })
