+++ test/multiSend.js
@@ -25,7 +25,7 @@ contract('MultiSend', function(accounts) {
         // Create Gnosis Safe and MultiSend library
         lw = await utils.createLightwallet()
         gnosisSafe = await utils.deployContract("deploying Gnosis Safe Mastercopy", GnosisSafe)
-        await gnosisSafe.setup([lw.accounts[0], lw.accounts[1]], 1, 0, 0, 0, 0, 0)
+        await gnosisSafe.setup([lw.accounts[0], lw.accounts[1]], 1, 0, 0, 0, 0, 0, 0)
         multiSend = await MultiSend.new()
         createAndAddModules = await CreateAndAddModules.new()
